requirements.txt
Flask==1.1.2
jsonschema==3.2.0
gunicorn==20.0.4
requests==2.24.0
requests-oauthlib==1.3.0
flask_bootstrap==3.3.7.1
python-dotenv==0.17.1
